Use prop-types instead of proptypes, and files reorganization (#152)
* Use prop-types instead of proptypes
* Reorganize client files
* One line per import to reduce diffs
* Fixes #135
* Merge two files
* Add a question to Train test to avoid proptypes error
* Correct typo in name
