+++ client/package.json
@@ -17,7 +17,7 @@
     "find-config": "^1.0.0",
     "node-sass": "^4.14.1",
     "path": "^0.12.7",
-    "proptypes": "^1.1.0",
+    "prop-types": "^15.7.2",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-git-info": "^2.0.0",
