+++ client/src/components/quiz/Answers.js
@@ -1,5 +1,5 @@
 import React from "react";
-import PropTypes from "proptypes";
+import PropTypes from "prop-types";
 
 /* ---------- Buttons ---------- */
