+++ client/src/components/quiz/Message.js
@@ -1,5 +1,5 @@
 import React from "react";
-import PropTypes from "proptypes";
+import PropTypes from "prop-types";
 
 const Message = ({ type, content }) => {
   let title;
