+++ client/src/components/quiz/Question.js
@@ -1,5 +1,5 @@
 import React from "react";
-import PropTypes from "proptypes";
+import PropTypes from "prop-types";
 
 const Question = ({ numero, text }) => {
   return (
