+++ client/src/components/system/InstallApp.js
@@ -1,5 +1,5 @@
 import React from "react";
-import PropTypes from "proptypes";
+import PropTypes from "prop-types";
 import { DownloadIcon } from "@modulz/radix-icons";
 
 const InstallApp = ({ installPromptEvent }) => (
