+++ client/src/pages/Login.js
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 
-import AuthService from "../../services/auth.service";
+import AuthService from "../services/auth.service";
 
 class Login extends Component {
   constructor() {
