+++ client/src/styles/base/_App.scss
@@ -1,33 +1,20 @@
 html,
-body {
+body,
+#root {
   font-family: sans-serif;
   height: 100%;
   width: 100%;
 }
 
-b {
-  font-weight: bold;
-}
-
-#root {
-  height: 100%;
-  width: 100%;
-}
-
-svg {
-  vertical-align: top;
-}
-
-h1 {
-  color: #252525;
-}
-
-p {
-  color: #4a4a4a;
-}
-
-summary {
-  cursor: pointer;
+/**
+ * Main contain the content of each page
+ */
+main {
+  padding: 0;
+  margin: 0;
+  min-height: calc(100% - 50px);
+  background-color: $bg;
+  box-sizing: border-box;
 }
 
 @media only screen and (min-device-width: $desktop-size) {
