+++ client/src/styles/base/_mixins.scss
@@ -1,6 +1,5 @@
 @mixin not-selectable {
   user-select: none;
-  -moz-user-select: none;
   -webkit-user-drag: none;
   -webkit-user-select: none;
   -ms-user-select: none;
