+++ client/src/styles/base/_variables.scss
@@ -10,6 +10,8 @@ $mascot-blue-dark: #2d779b;
 
 $bg: #ffffff;
 
+$text-color: #4a4a4a;
+
 /* ---------- Buttons ---------- */
 
 $button-bg: #1e2034;
