+++ client/src/styles/components/_TopBar.scss
@@ -7,45 +7,52 @@ nav {
   background-color: $top-bar-bg;
 
   grid-template-columns: 30% 40% 30%;
-}
 
-nav > * {
-  display: inline-block;
-  margin: auto;
-  color: $top-bar-fg;
-  @include not-selectable;
+  > * {
+    display: inline-block;
+    margin: auto;
+    color: $top-bar-fg;
+    @include not-selectable;
 
-  &:first-child {
-    margin-left: 5px;
+    &:first-child {
+      margin-left: 5px;
+    }
+
+    &:last-child {
+      margin-right: 5px;
+    }
   }
-
-  &:last-child {
-    margin-right: 5px;
-  }
-}
-
-nav a {
-  text-decoration: none;
-}
-
-nav h1 {
-  text-align: center;
 
   a {
+    text-decoration: none;
+  }
+
+  h1 a {
     text-align: center;
     text-transform: uppercase;
-    height: 16px;
     font-weight: bolder;
     padding: 10px 15px;
     border-radius: 10px;
     background-color: gray;
 
-    &:visited {
+    &:visited,
+    &:link {
       color: $top-bar-fg;
     }
   }
+
+  #offlineBanner div {
+    margin-left: auto;
+  }
 }
 
-nav #offlineBanner div {
-  margin-left: auto;
+#userBadge {
+  display: flex;
+
+  > span {
+    text-align: center;
+    vertical-align: center;
+    align-self: center;
+    margin: 5px;
+  }
 }
