--- client/src/styles/components/_userBadge.scss
@@ -1,10 +1,0 @@
-#userBadge {
-  display: flex;
-
-  > span {
-    text-align: center;
-    vertical-align: center;
-    align-self: center;
-    margin: 5px;
-  }
-}
