+++ client/src/styles/pages/_About.scss
@@ -5,16 +5,20 @@ main#informations {
   align-items: stretch;
   padding: 10px;
 
+  text-align: center;
+  line-height: 1.1;
+
   p {
-    text-align: center;
     margin: 1rem auto;
+  }
+
+  #credits,
+  #app-version {
+    margin: 0.5rem auto;
+    font-size: 0.9rem;
   }
 
   #credits {
     margin-top: auto;
-
-    ol {
-      margin-top: 1rem;
-    }
   }
 }
