+++ client/src/styles/pages/_Login.scss
@@ -1,4 +1,4 @@
-#login {
+main#login {
   display: flex;
   flex-direction: column;
   align-items: center;
